Sphinx/requirements.txt
colorama==0.4.6
yara-python==4.3.1
requests==2.31.0
beautifulsoup4==4.12.3
tldextract==5.1.1
pygments==2.18.0
capstone==5.0.1